LUassignment.py
Day 2 Assignment
Write a Python program to classify a given number into one of three categories: negative, zero, or positive. The program should take an integer input and print out one of the following statements based on the input:


"The number is negative." if the number is less than zero.
"The number is zero." if the number is exactly zero.
"The number is positive." if the number is more than zero.